feat: Add Battle System with Attack Speed (Week 2 Features 1-3)
- Feature 1: Monster Entity & Repository with 5 monster seeds
- Feature 2: Battle System Core Logic (Priority Queue event-driven)
- Feature 3: Battle Controller & API (auto-reward on victory)
- Feature 4: Add AttackSpeed to Character and Monster stats

+++ IdleRPG.Application/DTOs/Rewards/RewardDto.cs
@@ -6,14 +6,14 @@ namespace IdleRPG.Application.DTOs.Rewards
     public class RewardDto
     {
         /// <summary>
-        /// 획득 골드
+        /// 획득 골드 (long 타입: 방치형 게임 특성상 큰 숫자 처리)
         /// </summary>
-        public int Gold { get; set; }
+        public long Gold { get; set; }
 
         /// <summary>
-        /// 획득 경험치
+        /// 획득 경험치 (long 타입: 방치형 게임 특성상 큰 숫자 처리)
         /// </summary>
-        public int Experience { get; set; }
+        public long Experience { get; set; }
 
         // TODO: 나중에 아이템 보상, 스킬북 등 추가 가능
         // public List<ItemReward> Items { get; set; } = new();
